added username and password forms
