+++ src/screens/Home.js
@@ -1,8 +1,24 @@
 import React from 'react';
-import { StyleSheet, Text, View, Image, Button } from 'react-native';
+import { StyleSheet, Text, View, Image } from 'react-native';
 import logo from '../../assets/CorrectDoggo.png';
 import { StatusBar } from 'expo-status-bar';
+import { Container, Content, Header, From, Input, Item, Label, Form, Button, Icon } from 'native-base';
 
+// this gives us the capability of using all the firebase methods in our application
+import * as firebase from 'firebase';
+
+//initialize firebase storing all the information regarding our firebase application
+const firebaseConfig = {
+    apiKey: "AIzaSyDtvmX8T-1N5OBBXVij228d0QtWkuORYpg",
+    authDomain: "voffaland-2db0c.firebaseapp.com",
+    databaseURL: "https://voffaland-2db0c-default-rtdb.europe-west1.firebasedatabase.app",
+    projectId: "voffaland-2db0c",
+    storageBucket: "voffaland-2db0c.appspot.com",
+};
+
+// run the firebase initialize app with the firebase config then we can run the firebase functions that are
+// available in the SDK (SEEMS LIKE IT IS UNESESARY)
+// firebase.initializeApp(firebaseConfig);
 
 const Home = ({ navigation }) => {
     return (
@@ -10,9 +26,37 @@ const Home = ({ navigation }) => {
             <Image source={logo} style={styles.logo} />
             <Text style={styles.VFtext}>VoffaLand </Text>
             <Button
-                title="Skoða HundaLand án innskráningu"
-                onPress={() => navigation.navigate("Hundasvæði")} />
+                full
+                success
+                onPress={() => navigation.navigate("Hundasvæði")}>
+                <Text>Skoða VoffaLand án innskráningu</Text>
+            </Button>
             <StatusBar style="auto" />
+            <Container style={styles.container}>
+                <Form>
+                    <Item floatingLabel>
+                        <Label>Email</Label>
+                        <Input
+                            autoCorrect={false}
+                            autoCapitalize="none"
+                        />
+                    </Item>
+                    <Item floatingLabel>
+                        <Label>Password</Label>
+                        <Input
+                            secureTextEntry={true}
+                            autoCorrect={false}
+                            autoCapitalize="none"
+                        />
+                    </Item>
+                    <Button
+                        full
+                        rounded
+                        success>
+                        <Text>Login</Text>
+                    </Button>
+                </Form>
+            </Container>
         </View>
     );
 };
@@ -22,9 +66,9 @@ const styles = StyleSheet.create({
     container: {
         flex: 1,
         backgroundColor: '#fff',
-        alignItems: 'center',
+        // alignItems: 'center',
         justifyContent: 'center',
-        marginBottom: 200
+        // marginBottom: 200
     },
     logo: {
         width: 300,
